refac: Refactored CRUD methods
