+++ web_app/db/crud/__init__.py
@@ -1,0 +1,10 @@
+"""
+This module contains the CRUD operations for the database.
+"""
+
+from .airdrop import *
+from .base import *
+from .deposit import *
+from .position import *
+from .telegram import *
+from .user import *
